script to see PI overlap
R Script with the start of the work to check if PI is in the author list. Also CSplits authors out horizontally and begins edgelist creation for network analysis. Still need to do fuzzy match, or exact match with toLower() to check if same last name and first initial. Ambiguity is a problem, if the PI has same last name and first initial as a different authors with same last name and first intitial.